Add files via upload
My main goal for this project was to create a project that shows my skills in data cleaning. I downloaded a kaggle dataset about sales analyst for electronic products. I first started off creating a working sheet, added filters to each column. I then removed duplicates, used TRIM function to remove spaces, and deleted blank spaces. In addition I changed the formatting for columns that had a numerical values. I decided to split the address column to three different columns so that we could look at each states purchases. Last lastly, I created two additional columns for for the Price Each column. I added a discount rate, total price and a column for the updated price. I used the IF function to make an statement to add a discount if the total price was over $150. Then I created a IF function nested into another IF function to calculate the update total. FInally I created a dashboard with charts I made from pivot tables. 

Analysis:
After finishing the dashboard I concluded that there is trend between San Francisco and the largest purchases. San Francisco has made the largest purchase for electronic products. I also found that usb cables and lightning cables were the most purchased product. With this information we could use this to help inventory to keep with the demand for these products.